Keep it simple
Try to use fewer dependencies, get a standalone set of scripts that do one
thing: download data from a pump.

Copy cl2.py over to this repo, away from insulaudit, try to work out some
simpler scripts in order to make it simple for people to participate, as well
as do a better exposition of the problems we are trying to solve.

Edit out spurious stuff.


Former-commit-id: 8037a4a49731828d171cfa9df96363c78c6ce92c