Capture decent runs of the carelink protocol.
So others can start to contribute?  Unfortunately, runs are not super stable.
In the current form, in order to get a run like this, you need to uncommonet
the PowerControl block, run cl2.py, then comment the block, then unplug the
device, replug it in and then re-run cl2.py.  It's quite an ordeal... many
apologies.
The good news is that if you can get the stars and the planets to align just
right, this cl2.py script will download and decode most of the current settings
from your Minimed > 5XX insulin pump.
Furthermore, it will (attempt to) download all available pages of
history known to the pump's logs, using the ReadHistoryData command.
These logs contain all glucose, bolus, usage of the bolus wizard, and
many other medical records for the patient's analysis.

By preparing these log files with the tools provided, it's possible to
to analyze the files to confirm the observations found in Carelink's
CSV export.  The list_dates.py script attempts to parse these files
to identify individual records.  Comparing output with CSV makes
further decoding very very clear.

Currently, I've hopefully simplified the source tree so that no
dependency is required.

Many thanks to @kenthorvath, @ehwest, and @n-west for their help.


Former-commit-id: e83dbd79da680b9b74d3f25c3905cd732c9d5d32