capture experimental results
Looks like using a sync_read_write kind of function to ensure that
stick is paying attention to you is advisable and required.  It also
produces nice results.

The way this works now is to use the same logic for transacting stick
commands: really try hard to ensure that the message you've written
comes back with a response.

Carelink's logs makes it look like they always perform two writes,
indiscriminately, but perhaps this is ok too.


Former-commit-id: 33d2e8085175b8091a0732bc44fe79ea043b99f4