add discussion of special bolus
I think this and the midnight records are the only ones left.
As for these special bolus opcodes, 0x5c, I suspect that 0x5b is it's own
opcode, ~11 body, and that 0x01 is it's own opcode (analysis shows it
always parses cleanly), and 0x5c is also it's own opcode.

0x5x seems to always require special parsing, always read up until the
null byte?  The open question is whether or not there are two null
bytes, and how to know if that's the case?


Former-commit-id: cf15102826d57bab013bc39a5971ca9e051cdf91
Former-commit-id: 4d23fb6ac45a75f2b5ff745fab3114e6a3797f65 [formerly 93a82c30d4019ae83ee1245c42d13745873a8d59]
Former-commit-id: 9c382d8077b44cd186854710298a6cd8a355b60c
Former-commit-id: 0aeee9a05971c9f29de857f9dfe7248eae04b8bc