take rate and duration as commandline parameters
The duration is described in 30 minute intervals.
Eg, a duration of 0x01 is 30 minutes.

The rate is described in 0.025 increments, at least on a 515.
Eg, a rate of 0x01 is 0.025U/H.

Conversion for both goes like this:
+  duration = args.duration / 30
+  rate = int(args.rate / 0.025)

The temp basal command, is 76, eg 0x4C hex, and takes a parameters
list of 3 bytes.

I have no idea what the first byte does, this particular logic only
works when the first byte is set to 0.

This logic sets an absolute rate of units/hour, not a percentage,
there may or may not be another way to do that.

The second byte of the parameters represents the rate, and the third
is the duration.

The rate is an integer describing 30 minute increments.  Duration is
an integer describing 0.025 unit increments.

See --help for more on the --rate and --duration parameters, which
takes minutes, and absolute units per hour.


Former-commit-id: 18a5537d941292ba5378a029002eb5ffeabd101d
Former-commit-id: 184dfd6fb9dbcddcec2be2849dc91d2b3d40f3ef [formerly d7d1021c11ada658be70e4d1eeefd9012ee66d84]
Former-commit-id: 59d50d2d73d5570516fe944fd8b88a0c08b7f894
Former-commit-id: c3dfe18dc37629b070a4d039b5ac2ebd21ee65bf