allow some breathing room between fast polling loops
When sustaining many dozens/hundreds of commands all at once, the pump
sometimes needs a little breathing room.

According to the analysis, there should be some byte which describes
the need for the comms to pause for some time, ~3 seconds, and/or
until an ACK is recieved.  Instead, sometimes introduce longer delays
before attempting to continue, if it seems like communications have
hung.


Former-commit-id: b866d9c79d99c033349cd82350615980539875d7
Former-commit-id: 2bb9947d979c6b3bb6c87f24c30c18e135714351 [formerly 45faa706ded0d149fa5bd82318ad820d2dfcd9b7]
Former-commit-id: dc84d88f3ce7b414dcd3c2b0879a87b0a8cf73c8
Former-commit-id: fbede322fc5a59fd6083dbfbee1b88d981cf63c7