cgm download pages identical
When downloading the CGM data from the MM pump. The download command
seems to be downloading the same page over and over again. All the pages
downloaded (1-9) appear to be identical.

Could be that the data in them is identical and that the downloading
process is working correctly, but it just seems a little odd.


Former-commit-id: e936e5c331296fe0db2e00088b6d0f09d71b3348
Former-commit-id: d008000f74f53a2d7db84e678f394e6f21ff4a43 [formerly 8f774c3b5193fd89156223024dfef4cf5c9a5810]
Former-commit-id: a54e13eb5a04019e417eacbf1c506d763f01ecc1
Former-commit-id: 36edf8b729a3713bc5e5dc0f4c3550037b757ef8