cgm download pages identical
When downloading the CGM data from the MM pump. The download command
seems to be downloading the same page over and over again. All the pages
downloaded (1-9) appear to be identical.

Could be that the data in them is identical and that the downloading
process is working correctly, but it just seems a little odd.


Former-commit-id: 6ec2b94e01483e32695a3f6caf1e8688065fe310
Former-commit-id: 5586b9891290f47cae5be5882d365e877d7be31a [formerly aade01feee5327fc81a5d4c01a54630995223d8f]
Former-commit-id: 6c9f669b47a9072a32d21614017b8e9b34e6c1a9
Former-commit-id: 1ecaac53a50da1e050d9db6f1310e16d43518a8a