bolus: force explicit selection of stroke size
Make the bolus tool safer; the stroke sizes for pumps vary widely by model.
This script has only been tested on 515.

This change forces the user to explicitly declare either the pump
model or the stroke size to use.  This should allow easily re-using
the tool to discover stroke sizes fairly easily, in order to add the
pump specific option.

The tool will error out and refuse to do anything if the
strokes_per_unit has not been explicitly indicated, either through
specifying the pump model or the strokes option.


Former-commit-id: b8e2d72a721e7e7e54a943cf2b4e3c45168d656a
Former-commit-id: 4cefad78335f97effa4c80c284d6ab71733e275b [formerly a9620f424b02ec53c9efe1ecc311ad2441e2b2da]
Former-commit-id: 77330a1d63f1c55fe07832637bf01d13a216440e
Former-commit-id: 2fc0de0edd729d4346f9d65f7ea1266769db7a86