[mmeowlink_temp_basal] Handle empty byte arrays on temp basal response
In certain cases with mmeowlink, self.data is a blank bytearray as there is no
radio data returned. Attempting to access the 0th element raises 'IndexError:
bytearray index out of range' in this case.

This change adds a simple guard clause to avoid the exception. When the correct
data is actually set in the data array on the second attempt, the code works as
expected and the temporary basal report stores the correct data.


Former-commit-id: 2869cffb52979a9aacbf6b6da42155c3513f2e19
Former-commit-id: 7ac42fc7855f961f83f389a809536693e9712bb9 [formerly 788501fb39023a2472f38a9276e2514d88f426f4]
Former-commit-id: 5b1555b3a51ea800eea0078cd113fa0dbbf4e208
Former-commit-id: 77af8bc2d595ac7454c253b64f6c10428ac838d0